requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
Pillow==10.4.0
google-generativeai==0.8.3
requests==2.32.3
aiofiles==23.2.1
